Change logic for determining if devicetracker_id is set
Previously was looking to see if the devicetracker_id existed, was not None and state was not NotSet.

Sometimes person entities (and others) have a state of NotSet but do have lat/long.

Changed logic to remove state check of NotSet and instead see if devicetracker_id has latitude and longitude attributes.